Refactor: move internal replication request/response data to separate file
